Simplify npc message text
Made instruction easier to understand.
Need to fix problem with Remove from parent, when player interacts wit NPC and message appears on screen, it remains on screen for 4 seconds, if NPC interacts with player again during that time before the message disappears and a new message appears on screen, the previous messages is not removed and stays on screen.