Simplify npc Message
Made message easier to understand.
Need to fix problem with the NPC message when player interacts with NPC and message appears on screen, it has a delay of 4 seconds, if NPC interacts with player befor the message is removed from the parent anew message appears when player moves away the previous message stays on screen and doesn't get removed.